Storage interface
I renamed the SessionStorage as Storage and then make an implementation
of it.
The implementation "StateStorage" use as storage a redis database:
- Create: saves a model.Authorization in a redis database
- Obtain: gets a model.Authorization from redis key
- Delete: removes a key from redis if exists

Also I created unit tests for all methods of StateStorage.
Note: Unit tests have clean up functions to leave the database as it was before testing.
(Still they should not be tested in a production environment)

issue state-storage